refactor(store): organize store files
Organize store files; seperate reducers, effects and payloads into files

BREAKING CHANGE: Change file directory, new files added

re #5
