+++ libs/models/src/index.ts
@@ -1,6 +1,6 @@
-export * from './lib/data/player';
-export * from './lib/data/race';
-export * from './lib/data/tournament';
-export * from './lib/data/leaderboard';
-export * from './lib/data/stateModel';
-export * from './lib/data/errorLog';
+export * from './lib/state/player';
+export * from './lib/state/race';
+export * from './lib/state/tournament';
+export * from './lib/state/leaderboard';
+export * from './lib/state/stateModel';
+export * from './lib/state/errorLog';
