+++ libs/models/src/lib/state/stateModel.ts
