+++ libs/store/src/lib/effects/race.ts
@@ -1,12 +1,7 @@
 import {
   AppErrorCode,
-  AppErrorLog,
-  AppIdNumberType,
   AppPlayerId,
-  AppPlayerLog,
-  AppPlayerLogId,
   AppPlayerProfiles,
-  AppPlayerState,
   AppRace,
   AppRaceId,
   AppTournament,
@@ -14,25 +9,20 @@ import {
   AppTournamentState,
 } from '@razor/models';
 import {
-  generateUid,
-  generateAvatarLink,
-  giveZeroPadding,
   calculateTimeoutTimer,
-  generateLeaderboard,
   extractId,
   extractIdType,
+  generateLeaderboard,
+  giveZeroPadding,
 } from '@razor/util';
 import {
-  clearPlayerPayload,
   endCountdownPayload,
   endRacePayload,
-  joinPlayerPayload,
-  sendTypeLogPlayload,
-  setReadyTournamentPayload,
   startCountdownPayload,
-} from './payloadTypes';
-import { Dispatch, RootState } from './store';
+} from '../payloads';
+import { Dispatch, RootState } from '../store';
 
+//TODO: should move this to server later
 const loadRacingText = async (): Promise<string> => {
   const url = 'http://www.metaphorpsum.com/paragraphs/1/8';
 
@@ -41,89 +31,6 @@ const loadRacingText = async (): Promise<string> => {
     .then(data => {
       return data;
     });
-};
-
-//TODO: Add logger methods for warns and errors
-export const joinPlayer = async (
-  dispatch: Dispatch,
-  payload: joinPlayerPayload,
-  state: RootState,
-): Promise<void> => {
-  const { tid, playerName }: { tid: string; playerName: string } = payload;
-  let tournamentId: string;
-  if (tid) {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    if (!state.game.tournamentsModel[tid as any]) {
-      dispatch.game.sendErrorLog({
-        message: `Tournament with id ${tid} does not exist`,
-        code: AppErrorCode.TournamentNotExists,
-        relatedId: '',
-      });
-      return;
-    }
-    tournamentId = tid;
-  } else {
-    tournamentId = await generateUid(AppIdNumberType.Tournament);
-  }
-
-  const formattedTournamentId: AppTournamentId =
-    tournamentId as AppTournamentId;
-  const formattedPlayerId: AppPlayerId = (await generateUid(
-    AppIdNumberType.Player,
-  )) as AppPlayerId;
-
-  if (!tid) {
-    //Creating new tournament
-    dispatch.game.addTournamentReducer({
-      tournamentId: formattedTournamentId,
-      tournament: {
-        state: AppTournamentState.Lobby,
-        raceIds: [],
-        playerIds: [],
-      },
-    });
-  }
-  //Joining tournament
-  dispatch.game.addPlayerReducer({
-    tournamentId: formattedTournamentId,
-    playerId: formattedPlayerId,
-    player: {
-      name: playerName,
-      avatarLink: await generateAvatarLink(playerName),
-      state: AppPlayerState.Idle,
-      tournamentId: formattedTournamentId,
-    },
-  });
-};
-
-export const clearPlayer = async (
-  dispatch: Dispatch,
-  payload: clearPlayerPayload,
-  state: RootState,
-): Promise<void> => {
-  const { playerId }: { playerId: AppPlayerId } = payload;
-  const tournamentId = state.game.playersModel[playerId].tournamentId;
-  dispatch.game.removePlayerReducer({
-    tournamentId,
-    playerId,
-  });
-};
-
-export const setReadyTournament = async (
-  dispatch: Dispatch,
-  payload: setReadyTournamentPayload,
-  state: RootState,
-): Promise<void> => {
-  const { tournamentId }: { tournamentId: AppTournamentId } = payload;
-  const tournament: AppTournament = {
-    state: AppTournamentState.Ready,
-    raceIds: [...state.game.tournamentsModel[tournamentId].raceIds],
-    playerIds: [...state.game.tournamentsModel[tournamentId].playerIds],
-  };
-  dispatch.game.updateTournamentReducer({
-    tournamentId,
-    tournament,
-  });
 };
 
 export const startCountdown = async (
@@ -250,36 +157,5 @@ export const endRace = async (
   dispatch.game.updateLeaderboardReducer({
     leaderboardId: raceId,
     leaderboard,
-  });
-};
-
-export const sendTypeLog = async (
-  dispatch: Dispatch,
-  payload: sendTypeLogPlayload,
-): Promise<void> => {
-  const {
-    raceId,
-    playerId,
-    playerLog,
-  }: { raceId: AppRaceId; playerId: AppPlayerId; playerLog: AppPlayerLog } =
-    payload;
-
-  const playerLogId: AppPlayerLogId = `${raceId}-${playerId}`;
-
-  dispatch.game.updatePlayerLogReducer({
-    playerLogId,
-    playerLog,
-  });
-};
-
-export const sendErrorLog = async (
-  dispatch: Dispatch,
-  payload: AppErrorLog,
-): Promise<void> => {
-  const errorTimestamp: number = new Date().getTime();
-  const errorLog = payload;
-  dispatch.game.logErrorReducer({
-    errorLog,
-    errorTimestamp,
   });
 };
