+++ libs/store/src/lib/effects/tournament.ts
@@ -1,0 +1,24 @@
+import { setReadyTournamentPayload } from '../payloads';
+import { Dispatch, RootState } from '../store';
+import {
+  AppTournament,
+  AppTournamentId,
+  AppTournamentState,
+} from '@razor/models';
+
+export const setReadyTournament = async (
+  dispatch: Dispatch,
+  payload: setReadyTournamentPayload,
+  state: RootState,
+): Promise<void> => {
+  const { tournamentId }: { tournamentId: AppTournamentId } = payload;
+  const tournament: AppTournament = {
+    state: AppTournamentState.Ready,
+    raceIds: [...state.game.tournamentsModel[tournamentId].raceIds],
+    playerIds: [...state.game.tournamentsModel[tournamentId].playerIds],
+  };
+  dispatch.game.updateTournamentReducer({
+    tournamentId,
+    tournament,
+  });
+};
