+++ libs/store/src/lib/payloads/reducers.ts
@@ -1,18 +1,18 @@
 import {
+  AppTournamentId,
+  AppTournament,
+  AppRaceId,
+  AppRace,
+  AppPlayerId,
+  AppPlayer,
+  AppLeaderboard,
+  AppPlayerLogId,
+  AppPlayerLog,
   AppErrorLog,
   AppErrorTimestamp,
-  AppLeaderboard,
-  AppPlayer,
-  AppPlayerId,
-  AppPlayerLog,
-  AppPlayerLogId,
-  AppRace,
-  AppRaceId,
-  AppTournament,
-  AppTournamentId,
 } from '@razor/models';
 
-//reducerPlayload
+//reducer Playloads
 export type addTournamentReducerPayload = {
   tournamentId: AppTournamentId;
   tournament: AppTournament;
@@ -52,32 +52,4 @@ export type removeTournamentReducer = {
 export type logErrorReducerPayload = {
   errorLog: AppErrorLog;
   errorTimestamp: AppErrorTimestamp;
-};
-
-//effectsPayloads
-export type joinPlayerPayload = {
-  tid: string;
-  playerName: string;
-};
-export type clearPlayerPayload = {
-  playerId: AppPlayerId;
-};
-export type setReadyTournamentPayload = {
-  tournamentId: AppTournamentId;
-};
-export type startCountdownPayload = {
-  tournamentId: AppTournamentId;
-  playerId: AppPlayerId;
-};
-export type endCountdownPayload = {
-  tournamentId: AppTournamentId;
-};
-export type endRacePayload = {
-  raceId: AppRaceId;
-};
-export type sendTypeLogPlayload = {
-  raceId: AppRaceId;
-  playerId: AppPlayerId;
-  /** Timestamp, and text length from players machine */
-  playerLog: AppPlayerLog;
 };
